requirements.txt
flask
requests
beautifulsoup4
gunicorn
pymongo[srv]
python-dotenv